requirements.txt
environs~=14.1.1
openai~=1.79.0
Django>=5.2.1,<5.3
djangorestframework>=3.16.0,<3.17
pytest>=8.3.5,<8.4
pytest-django>=4.11.1,<4.12
celery>=5.5.2,<5.6
redis>=6.1.0,<6.2
drf-spectacular>=0.28.0,<0.29
drf-spectacular-sidecar>=2025.5.1,<2025.6
django-filter>=25.1.0,<25.2